# Add support for compact and masked watersheds