This PR adds support for the compact watershed algorithm (https://www.tu-chemnitz.de/etit/proaut/publications/cws_pSLIC_ICPR.pdf). This algorithm is nice because it results in more uniform watershed basins as you increase the `compactness` parameter. 

I still need to do the following before I think it should be merged:

- [x] Add tests
- [x] More docs (and reference!)
- [x] Benchmarking (thanks to https://github.com/scikit-image/scikit-image/issues/2636 for speed enhancements)

Example segmentation following the coin example from the docs:

| Simple Watershed  | Compact watershed  |
|---|---|
| ![simple](https://user-images.githubusercontent.com/1661487/57992421-e6dde500-7a68-11e9-8ee5-3111fbf73a53.png) | ![compact](https://user-images.githubusercontent.com/1661487/57992418-e1809a80-7a68-11e9-955d-7a58e5dba775.png) |


closes #34