@tlnagy 2019-05-20
Looks like my PR introduces quite a few allocations which are causing some slowdowns:

Master

```julia
julia> @benchmark segments = watershed(dist, markers)
BenchmarkTools.Trial: 
  memory estimate:  3.24 MiB
  allocs estimate:  119
  --------------
  minimum time:     41.722 ms (0.00% GC)
  median time:      42.512 ms (0.00% GC)
  mean time:        43.070 ms (0.92% GC)
  maximum time:     48.598 ms (5.49% GC)
  --------------
  samples:          117
  evals/sample:     1
```

This PR

```julia
julia> @benchmark segments = watershed(dist, markers)
BenchmarkTools.Trial: 
  memory estimate:  21.91 MiB
  allocs estimate:  581103
  --------------
  minimum time:     102.520 ms (3.66% GC)
  median time:      103.819 ms (3.65% GC)
  mean time:        105.706 ms (3.63% GC)
  maximum time:     152.963 ms (3.43% GC)
  --------------
  samples:          48
  evals/sample:     1
```