@tlnagy 2020-02-06
I rebased, fixed the PR based on your suggestions, fixed a representation error bug with segmenting fixed point number images, and added tests. 

I think this is ready to merge once the tests pass.