@tlnagy 2020-02-06
@timholy Whenever you get a chance, could you merge this and tag a new release? Thanks!