@johnnychen94 2020-02-11
Reminder: retrigger the test after https://github.com/JuliaImages/ImageSegmentation.jl/pull/44 being merged.