@tlnagy 2020-04-08
I updated this PR by adding back the type paramerization in PixelKey which fixes the performance regressions, it's basically back to the original speed. The compact algorithm now always uses floats to avoid the representation error when accumulating using fixed point numbers. @timholy and @johnnychen94, I believe this is ready to merge.

This PR:

```julia
julia> @benchmark segments = watershed(dist, markers)
BenchmarkTools.Trial: 
  memory estimate:  3.44 MiB
  allocs estimate:  121
  --------------
  minimum time:     35.417 ms (0.00% GC)
  median time:      39.968 ms (0.00% GC)
  mean time:        39.711 ms (0.11% GC)
  maximum time:     42.989 ms (0.00% GC)
  --------------
  samples:          126
  evals/sample:     1
```

Master:

```julia
julia> @benchmark segments = watershed(dist, markers)
BenchmarkTools.Trial: 
  memory estimate:  3.24 MiB
  allocs estimate:  119
  --------------
  minimum time:     33.161 ms (0.00% GC)
  median time:      36.747 ms (0.00% GC)
  mean time:        36.770 ms (0.11% GC)
  maximum time:     41.609 ms (0.00% GC)
  --------------
  samples:          136
  evals/sample:     1
```