@tlnagy 2020-04-08
Not sure why Travis isn't showing up under checks, but the build passed: https://travis-ci.org/github/JuliaImages/ImageSegmentation.jl/builds/672668131