@tlnagy 2020-06-10
@johnnychen94 and @timholy I finally got around to making all the suggested changes. Hope this is finally ready for a merge!