@tlnagy 2020-06-17
Could you tag a new version when you get a chance? Thanks!